Crate iceberg_playground

Introduction

Iceberg Playground (iceberg-playground) is a small command line utility that runs SQL queries against tables, which is backed by the DataFusion engine.

Structs

IcebergCatalogList

Constants

ICEBERG_PLAYGROUND_VERSION